requirements.txt
sqlalchemy
psycopg2
pytest
assertpy
logging
allure-pytest